--- src/View/__mocks__/utils.ts
@@ -1,2 +1,0 @@
-export const getView = jest.fn();
-export const listViews = jest.fn();
